// src/lib/cb-tags/types.ts
export type CourseBuilderTag = {
  id: string
  name: string
  slug: string
  label: string | null
  description: string | null
  image_url: string | null
  image_480_url: string | null
  path: string
  url: string | null
  context: string | null
  taggingsCount: number | null
  popularityOrder: number | null
  legacyId: number | null
}